src/app/requirements.txt
flask
google-cloud-secret-manager

